+++ android-pdf-viewer/src/main/java/com/github/barteksc/pdfviewer/annotation/ocg/OCGParser.java
@@ -1,4 +1,4 @@
-package com.github.barteksc.pdfviewer.annotation;
+package com.github.barteksc.pdfviewer.annotation.ocg;
 
 import com.lowagie.text.ExceptionConverter;
 import com.lowagie.text.pdf.PRStream;
