+++ sample/src/main/java/com/github/barteksc/sample/PDFViewActivity.java
@@ -40,7 +40,7 @@ import androidx.core.app.ActivityCompat;
 import androidx.core.content.ContextCompat;
 
 import com.github.barteksc.pdfviewer.PDFView;
-import com.github.barteksc.pdfviewer.annotation.OCGRemover;
+import com.github.barteksc.pdfviewer.annotation.ocg.OCGRemover;
 import com.github.barteksc.pdfviewer.link.LinkHandler;
 import com.github.barteksc.pdfviewer.listener.OnErrorListener;
 import com.github.barteksc.pdfviewer.listener.OnLoadCompleteListener;
